+++ src/components/LoginPage/index.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import Logo from '../../Images/logo-future-eats-invert.svg'
+import Logo from '../../Images/logo-future-eats@3x.png'
 import {ContainerLogin, LogoImg, TituloLogin, ContainerTitulo, StyledTextField, ContainerInputs, StyledButton, SignUpTitle, SignUpButton} from './styles'
 import InputAdornment from '@material-ui/core/InputAdornment';
 import IconButton from '@material-ui/core/IconButton';
